+++ src/App.js
@@ -1,35 +1,6 @@
 import React from 'react';
+import AddTodoForm from './AddTodoForm';
+import TodoList from './TodoList';
 
-const todoList = [
-  {
-    objectID: 0,
-    title: 'Finish lesson 1.1',
-  },
-  {
-    objectID: 1,
-    title: 'Upload to GitHub',
-  },
-  {
-    objectID: 2,
-    title: 'Turn in assignment with Squibbly',
-  }
-];
 
-function App() {
-  return (
-    <div>
-      <h1>Todo List</h1>
-      <ul>
-          {todoList.map(function (listItem) {
-            return (
-            <li key={listItem.objectID}>
-              <span>{listItem.title}</span>
-              </li>
-            );
-          })}
-      </ul>
-    </div>
-  );
-}
-
-export default App;
+export default TodoList;
